1.添加布隆过滤器 2.添加mybatis 分页插件 3.更新上传文件 4.添加关注与被关注 5.更新商品管理 6.添加用户属性 7.添加注册接口

+++ secondhand/src/main/java/com/secondhand/module/product/entity/Graph.java
@@ -23,13 +23,13 @@ public class Graph {
      * 用户id
      */
     @TableField(value = "uid")
-    private Integer uid;
+    private Long uid;
 
     /**
      * 粉丝id
      */
     @TableField(value = "followid")
-    private Integer followid;
+    private Long followid;
 
     public static final String COL_ID = "id";
